requirements.txt
cycler==0.10.0
joblib==0.14.1
kiwisolver==1.1.0
matplotlib==3.1.2
numpy==1.17.4
pandas==0.25.3
pyparsing==2.4.5
python-dateutil==2.8.1
pytz==2019.3
scikit-surprise==1.1.0
scipy==1.3.3
six==1.13.0
tabulate==0.8.6